requirements.txt
Flask
Flask-Cors
mysql-connector-python
face-recognition
opencv-python-headless
numpy
python-dotenv
PyJWT
bcrypt
